requirements.txt
setuptools>=49.2.1
discord.py
git+https://gitlab.com/Tmpod/libneko@commands-fix
requests
aiofiles
Pillow
async_tio
numpy
